feat : Shimmer를 Modifier 확장함수로 변경
