+++ feature/setting/src/commonMain/kotlin/com/whatever/caramel/feature/setting/component/UserProfile.kt
@@ -1,10 +1,5 @@
 package com.whatever.caramel.feature.setting.component
 
-import androidx.compose.animation.core.RepeatMode
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
@@ -19,10 +14,9 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
+import com.whatever.caramel.core.designsystem.components.shimmer
 import com.whatever.caramel.core.designsystem.foundations.Resources
 import com.whatever.caramel.core.designsystem.themes.CaramelTheme
 import com.whatever.caramel.core.domain.vo.user.Gender
@@ -106,25 +100,6 @@ internal fun SettingUserProfile(
 internal fun SettingUserProfileSkeleton(
     modifier: Modifier = Modifier
 ) {
-    val transition = rememberInfiniteTransition()
-    val translateAnim = transition.animateFloat(
-        initialValue = 0f,
-        targetValue = 1000f,
-        animationSpec = infiniteRepeatable(
-            animation = tween(durationMillis = 1200, delayMillis = 300),
-            repeatMode = RepeatMode.Restart
-        )
-    )
-    val shimmerColors = listOf(
-        CaramelTheme.color.skeleton.primary,
-        CaramelTheme.color.skeleton.secondary,
-        CaramelTheme.color.skeleton.primary
-    )
-    val brush = Brush.linearGradient(
-        colors = shimmerColors,
-        start = Offset.Zero,
-        end = Offset(x = translateAnim.value, y = translateAnim.value),
-    )
     Row(
         modifier = modifier,
         verticalAlignment = Alignment.CenterVertically
@@ -132,10 +107,7 @@ internal fun SettingUserProfileSkeleton(
         Box(
             modifier = Modifier
                 .size(50.dp)
-                .background(
-                    brush = brush,
-                    shape = CaramelTheme.shape.l
-                )
+                .shimmer(shape = CaramelTheme.shape.l)
         )
         Spacer(modifier = Modifier.padding(end = CaramelTheme.spacing.m))
         Column(
@@ -146,19 +118,13 @@ internal fun SettingUserProfileSkeleton(
             Box(
                 modifier = Modifier
                     .size(width = 100.dp, height = 22.dp)
-                    .background(
-                        brush = brush,
-                        shape = CaramelTheme.shape.xs
-                    )
+                    .shimmer(shape = CaramelTheme.shape.xs)
             )
             Spacer(modifier = Modifier.padding(bottom = CaramelTheme.spacing.xs))
             Box(
                 modifier = Modifier
                     .size(width = 70.dp, height = 16.dp)
-                    .background(
-                        brush = brush,
-                        shape = CaramelTheme.shape.xs
-                    )
+                    .shimmer(shape = CaramelTheme.shape.xs)
             )
         }
     }
